feature/SCIHUB-1209: whole modal as prop
